pyproject.toml
[project]
name = "app-runner-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "pyramid==2.0",
    "snowflake-connector-python>=4.2.0",
    "streamlit>=1.52.2",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "watchdog>=6.0.0",
]
